refactor: extract the interfaces to a single file
By having them in the same file it gives a overview of what the package
is about.
